+++ spells/index.js
@@ -1,33 +1,32 @@
 /* eslint-disable linebreak-style */
 // This is where we `require` all of our spell handlers ('event:name': handler)
-const lockSpells = require('./locks.js');
-const unlockStationSpell = require('./unlock-station.js');´
-const unlockPlayer = require('./unlock_player.js');
-const immuneToSpells = require('./immune_To_Spells');
-const swapTeamRack = require('./swap-team-rack');
-const halfSalaryStation = require('./half_salary_station');
-const doubleSalary = require('./double_salary.js');
-const doubleSalaryStation = require('./double_salary_station');
-const activateTemporaryPocket = require('./activate-temporary-pocket');
-const increaseLoginTimePlayer = require('./increase-login-time-player');
-const lockPlayer = require('./lock_player.js');¨
-const halfSalaryAllStations = require('./half-salary-all-stations');
+const playerImmune = require('./player-immune');
+const playerIncLogin = require('./player-increased-login');´
+const playerLocked = require('./player-locked');
+const playerTemporaryPocket = require('./player-temporary-pocket');
+const playerUnlocked = require('./player-unlocked');
 const resetAll = require('./reset-all');
-const temporaryPocket = require('./temporary-pocket');
+const stationDoubleSalary = require('./station-double-salary');
+const stationHalfSalary = require('./station-half-salary');
+const stationLock = require('./station-lock');
+const stationUnlock = require('./station-unlock');
+const stationsDoubleSalary = require('./stations-double-salary');¨
+const stationsHalfSalary = require('./stations-half-salary');
+const teamsSwapRack = require('./teams-swap-rack');
 
 // And then combine them into one object
 module.exports = {
-  lockSpells,
-  unlockStationSpell,
-  unlockPlayer,
-  immuneToSpells,
-  swapTeamRack,
-  halfSalaryStation,
-  doubleSalary,
-  doubleSalaryStation,
-  activateTemporaryPocket,
-  increaseLoginTimePlayer,
-  lockPlayer,
-  halfSalaryAllStations,
-  temporaryPocket,
+  playerImmune,
+  playerIncLogin,
+  playerLocked,
+  playerTemporaryPocket,
+  playerUnlocked,
+  resetAll,
+  stationDoubleSalary,
+  stationHalfSalary,
+  stationLock,
+  stationUnlock,
+  stationsDoubleSalary,
+  stationsHalfSalary,
+  teamsSwapRack,
 };
