+++ spells/player-increased-login.js
