+++ spells/player-locked.js
